+++ SecurityFever/Functions/SystemAudit/Get-SystemAuditWindowsService.ps1
@@ -1,0 +1,26 @@
+<#
+
+7045
+
+Im System wurde ein Dienst installiert.
+
+Dienstname:  Pml Driver HPZ12
+Dienstdateiname:  %SystemRoot%\System32\svchost.exe -k HPZ12
+Diensttyp:  Benutzermodusdienst
+Dienststarttyp:  Automatisch starten
+Dienstkonto:  NT AUTHORITY\LocalService
+
+Properties:
+ServiceName Pml Driver HPZ12
+ImagePath %SystemRoot%\System32\svchost.exe -k HPZ12
+ServiceType Benutzermodusdienst
+StartType Automatisch starten
+AccountName NT AUTHORITY\LocalService
+
+#>
+
+
+
+
+
+