+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Added: System audit functions
+* Added: TOTP Generator
+
 ## 2.8.1 - 2019-11-08
 
 * Fixed: Spelling of Add-CertificatePrivateKeyPermission command